Fixes #1159: catches telemetry lifecycle crash
RCA:
- Some devices are calling MainActivity#onPause before MainActivity#onResume
- This leads to an InvalidStateException from our Telemetry library
